Add forecast querying to the project.
This change adds forecaster.go, which will be used to query the openweathermap
API for weather forecast details. The next step will be properly parsing the
response from the GET request into datastructures that can be worked with.

This change also does some refactoring of mailer.go, pulling a bunch of stuff
out of the main() function to make it more modular. The end goal is for it to
simply be a library (this is true of the forecaster as well). I'll have a single
driver to tie them together.