Create the email contents.
Implement the CreateMessage function, which creates the subject and body of the
email that will be sent to the recipient.

This commit also contains some experimentation with storing the weather types in
proto format, one category per weather-type, with attached folksy sayings. I'm
still trying to figure out how I want to structure this, however.