Add protocol buffer for structured data.
This change adds a folder called proto/ which contains a proto definition for
storing structured weather+saying data, as well as a generated protobuf library
to manipulate it. It also includes a new file, generator.go, which can create a
protobuf with the weather+saying data for consumption by forecaster.go.