Finish off code to create the email message.
CreateMessage() will create the subject and body of the email message. It needed
a function to pull out the most commonly occurring weather type from the API
response, as well as the corresponding description, in order to create a
message for the user that would indicate the predominant form of weather that
would be encountered. The folksy saying picker also needs this information, so
that an appropriate saying could be picked for the predominate weather
condition.